Update compiler settings, fix couch_icu_driver.m
Set the deployment target to iOS 4.0.
Set optimization to -O3.
Remove unnecessary ICU headers.

Did NOT switch compiler from GCC to to LLVM-GCC yet -- we want the official builds to still use GCC for the time being because of the performance boost. This means anyone who has Xcode 4.2b6+ installed will first have to switch the compiler to LLVM-GCC in the iErl14 project to be able to build!

Change-Id: I2dfa0306d17600029cfcecf798ada68ceb63782c
Reviewed-on: http://review.couchbase.org/9171
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>