Avoid JS assertion failure on wake from sleep
emonk_main.c calls JS_SetCStringsAreUTF8 when registering monk. But that function has an assertion making sure that it can only ever be called once. To avoid this, just #define JS_C_STRINGS_ARE_UTF8 so we don't have to set this at runtime. It probably makes some string processing code faster too.
The same change to js-config.h has to be made to the copy of the file in iMonkey -- I'll check that in too.

Change-Id: I1091bf75115cc82102f7ae89cd33a6a092d092e4
Reviewed-on: http://review.couchbase.org/9527
Tested-by: Aaron Miller <apage43@ninjawhale.com>
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>
Tested-by: Jens Alfke <jens@couchbase.com>